+++ lib/defra_ruby/address/services/ea_address_facade_v1_service.rb
@@ -4,40 +4,11 @@ module DefraRuby
   module Address
     class EaAddressFacadeV1Service < BaseService
 
+      PARAM_NAME = "postcode"
+
       def run(postcode)
-        @postcode = postcode
-        Response.new(response_exe)
-      end
-
-      private
-
-      attr_reader :postcode
-
-      def url
-        host = DefraRuby::Address.configuration.host
-        client_id = DefraRuby::Address.configuration.client_id
-        key = DefraRuby::Address.configuration.key
-
-        File.join(
-          host,
-          "/address-service/v1/addresses/",
-          "postcode?client-id=#{client_id}&key=#{key}&postcode=#{postcode}"
-        )
-      end
-
-      def response_exe
-        lambda do
-          response = RestClient::Request.execute(
-            method: :get,
-            url: url,
-            timeout: DefraRuby::Address.configuration.timeout
-          )
-          results = JSON.parse(response)["results"]
-
-          raise DefraRuby::Address::NoMatchError if results.empty?
-
-          results
-        end
+        request = EaAddressFacadeRequest.new(PARAM_NAME)
+        request.execute(postcode)
       end
 
     end
